Move the C++ VFS out of common
The VFS interface is LLVM-specific and shouldn't
be visible external to the C++ indexer.

Partially addresses #112.
