+++ kythe/cxx/indexer/cxx/IndexerFrontendAction.cc
@@ -25,9 +25,9 @@
 #include "clang/Tooling/Tooling.h"
 #include "kythe/cxx/common/indexing/KytheClaimClient.h"
 #include "kythe/cxx/common/indexing/KytheGraphRecorder.h"
-#include "kythe/cxx/common/indexing/KytheVFS.h"
 #include "kythe/cxx/common/json_proto.h"
 #include "kythe/cxx/common/proto_conversions.h"
+#include "kythe/cxx/indexer/cxx/KytheVFS.h"
 #include "kythe/proto/analysis.pb.h"
 #include "kythe/proto/cxx.pb.h"
 #include "llvm/ADT/Twine.h"
