+++ kythe/cxx/indexer/cxx/KytheGraphObserver.h
@@ -28,10 +28,10 @@
 #include "GraphObserver.h"
 #include "kythe/cxx/common/indexing/KytheClaimClient.h"
 #include "kythe/cxx/common/indexing/KytheGraphRecorder.h"
-#include "kythe/cxx/common/indexing/KytheVFS.h"
 #include "kythe/cxx/common/kythe_metadata_file.h"
 #include "kythe/cxx/common/language.h"
 #include "kythe/cxx/indexer/cxx/IndexerASTHooks.h"
+#include "kythe/cxx/indexer/cxx/KytheVFS.h"
 #include "kythe/proto/storage.pb.h"
 
 namespace kythe {
