+++ kythe/cxx/indexer/cxx/KytheVFS.cc
