+++ kythe/cxx/indexer/cxx/KytheVFS.h
